Remove the need for config to be on PYTHONPATH.
This commit is actually a bit deeper than that, but that's the main effect.

Instead of changing sys.path to include the settings module, it is now
retrieved with imp and the normal import machinery is bypassed, making it
available in the interpreter without any special trickery.

See also the documentation in pydra/config.py for more information and usage.